% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{BuildDesign}
\alias{BuildDesign}
\title{Creates a design matrix for the experiment}
\usage{
BuildDesign(experimentNames, techReps, corrConst = NA)
}
\arguments{
\item{experimentNames}{Vector with names of the experiments}

\item{techReps}{Vector with number of technical replicates in each experiment}

\item{corrConst}{Optional, Vector with correction constants for each experiment}
}
\value{
Dataframe with experiments numbered and numbers of columns
}
\description{
Creates a design matrix for the experiment
}
\examples{

}
