% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeAICIs.R
\name{ComputeAICIs}
\alias{ComputeAICIs}
\title{ComputeAICIs}
\usage{
ComputeAICIs(
  inDF,
  vectReps,
  vectRepsCombsCC,
  pt = 0.5,
  Q = 0.95,
  BF = T,
  thr = NA,
  thrUP = NA,
  thrType = "each"
)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{vectRepsCombsCC}{A vector of pairwise-computed correction constants for given replicates}

\item{pt}{A value or a vector of values to compare with (if second, should be compatible with the order and size of genes vector in table of allelic counts)}

\item{Q}{An optional parameter; quantile (for example 0.95, 0.8, etc)}

\item{BF}{Bonferroni correction, set False if Q is alredy corrected}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
A table of gene names, AI, coverage, test p-value, and Confidence Intervals
}
\description{
Calculate Binomial and QCC-corrected binomial CIs for a given vector of point estimates, for given QCC.
}
