% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{ComputeCorrConstantFor2Reps}
\alias{ComputeCorrConstantFor2Reps}
\title{Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for each condition}
\usage{
ComputeCorrConstantFor2Reps(inDF, reps, binNObs = 40, EPS = 1.3,
  thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{reps}{A vector of 2 of replicate numbers that should be considered}

\item{binNObs}{Threshold on number of observations per bin}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
A table of gene names, AIs + CIs for each condition, classification into genes demonstrating differential AI and those that don't
}
\description{
Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for each condition
}
\examples{

}
