% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCorrConstantsForAllPairsReps.R
\name{ComputeCorrConstantsForAllPairsReps}
\alias{ComputeCorrConstantsForAllPairsReps}
\title{ComputeCorrConstantsForAllPairsReps}
\usage{
ComputeCorrConstantsForAllPairsReps(
  inDF,
  vectReps,
  binNObs = 40,
  fitCovThr = 50,
  EPS = 1.05,
  thr = NA,
  thrUP = NA,
  thrType = "each"
)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{binNObs}{Threshold on number of observations per bin}

\item{fitCovThr}{Threshold on coverage for genes that will be included in Beta-Bin fitting}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
List of fitting outputs of ComputeCorrConstantFor2Reps() for each combinatorial pair of replicates (in order 1-2,1-3,..,1-N,2-3,..2-N,..,(N-1)-N): list with (1) fitted QCC ($fittedCC) and (2) a table with proportions of observed to expected quantiles per coverage bin ($QObsExpPropsTable).
}
\description{
Computes QCC for all possible pairs of given replicates.
}
