% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsToAI.R
\name{CountsToAI}
\alias{CountsToAI}
\title{CountsToAI}
\usage{
CountsToAI(
  df,
  reps = NA,
  meth = "mergedToProportion",
  thr = NA,
  thrUP = NA,
  thrType = "each"
)
}
\arguments{
\item{df}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{reps}{Optional (default=NA, all replicates), a vector of replicate numbers for which the analysis should be applied}

\item{meth}{Optional (default="mergedToProportion", also can be "meanOfProportions"), method to use, either sum(m)/sum(p) (default), or sum(m/p)}

\item{thr}{Optional (default=NA), threshold on the overall number of counts for a gene to be considered in the analysis}

\item{thrUP}{Optional (default=NA), threshold for max gene coverage (default = NA)}

\item{thrType}{Optional (default = "each", also can be "average" for average coverage on replicates), threshold type}
}
\value{
A table with IDs and calculated AI estimate for given set of replicates
}
\description{
Calculates allelic imbalances from merged counts over given replicates (ai(sum_reps(gene))).
}
\examples{
CountsToAI(allelicCountsTable, reps=c(1,2), thr=10, thrUP=1000)

}
