% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{CreateCIforAI}
\alias{CreateCIforAI}
\title{Input: two data frames, one including replicate-gene mat|pat coverages, one with constants for each technical replicates pair}
\usage{
CreateCIforAI(dfInt, dfCounts, condName = "Condition", thr = NA,
  thrUP = NA, thrType = "each")
}
\arguments{
\item{dfInt}{A table with column "linInt" of correction constants for each replicates combination (rows), the order of rows should be consistent with columns in dfCounts, s.t. pairs are alphabetically ordered}

\item{dfCounts}{A table with pairs of columns for each technical replicate, the rows correspond to genes, the values are mat|pat coverages}

\item{condName}{An optional parameter; one-word name for condition}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
5-column df: ID, AI, Plus-minus intervals, AI Confidence Intervals left and right ends
}
\description{
Input: two data frames, one including replicate-gene mat|pat coverages, one with constants for each technical replicates pair
}
\examples{

}
