% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{CreateDeltaAIForAPairRepsDF}
\alias{CreateDeltaAIForAPairRepsDF}
\title{Creates a tab with gene mean coverage and AI deltas for a pair of technical replicates}
\usage{
CreateDeltaAIForAPairRepsDF(df, reps, thr = NA, thrUP = NA,
  thrType = "each")
}
\arguments{
\item{reps}{A parameter for setting 2 replicates for consideration}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{tab}{A dataframe of genes/transcripts and parental counts for technical replicates in columns}
}
\value{
A tab with names, gene mean coverage and AI deltas for a pair of technical replicates
}
\description{
Creates a tab with gene mean coverage and AI deltas for a pair of technical replicates
}
\examples{

}
