% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{CreateMergedDeltaAIPairwiseDF}
\alias{CreateMergedDeltaAIPairwiseDF}
\title{Creates a table of parwise comparisons for technical replicates in given table}
\usage{
CreateMergedDeltaAIPairwiseDF(df, reps = NA, what = "noname",
  thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df)}

\item{what}{A name (default = "noname")}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
A table of parwise comparisons for technical replicates in given table
}
\description{
Creates a table of parwise comparisons for technical replicates in given table
}
\examples{

}
