% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{CreateObservedQuantilesDF}
\alias{CreateObservedQuantilesDF}
\title{Creates a table with quantiles and numbers of bins for technical replicates for a given table, binned into log intervals}
\usage{
CreateObservedQuantilesDF(df, P, ep, logbase = T, coverageLimit,
  group = "")
}
\arguments{
\item{df}{A dataframe - output of CreateMergedDeltaAIPairwiseDF()}

\item{P}{A vector of %-quartiles}

\item{ep}{The log base for binning (0^b, 1^b, ...)}

\item{logbase}{The binary parameter if we deal with log base (default = T)}

\item{coverageLimit}{Gene coverage limit for consideration}

\item{group}{An optional name (default = '')}
}
\value{
A table with quantiles and numbers of bins for technical replicates for a given table
}
\description{
Creates a table with quantiles and numbers of bins for technical replicates for a given table, binned into log intervals
}
\examples{

}
