% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{CreatePMforAI}
\alias{CreatePMforAI}
\title{Input: two data frames, one including replicate-gene coverages, one with constants for each technical replicates pair}
\usage{
CreatePMforAI(dfInt, dfAI, dfCov)
}
\arguments{
\item{dfInt}{A table with column "linInt" of correction constants for each replicates combination (rows), the order of rows should be consistent with columns in dfCov, s.t. pairs are alphabetically ordered}

\item{dfAI}{A table with columns for each technical replicate, the rows correspond to genes, the values are AI}

\item{dfCov}{A table with columns for each technical replicate, the rows correspond to genes, the values are coverage}
}
\value{
Plus-minus intervals to determine AI Confidence Intervals for each gene
}
\description{
Input: two data frames, one including replicate-gene coverages, one with constants for each technical replicates pair
}
\examples{

}
