% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamASE.R
\name{DiffAIplusFig}
\alias{DiffAIplusFig}
\title{A function that gets a table with ref & alt counts per gene/SNP for each replicate}
\usage{
DiffAIplusFig(inDF, CondReps_expA, CondReps_expB, CondReps_expA_name,
  CondReps_expB_name, expA_CC = NA, expB_CC = NA, thr_coverage = NA,
  minDifference = 0.1, lm = T, save = F)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{CondReps_expA}{A vector (>=2) of replicate numbers that should be considered as first condition's tech reps}

\item{CondReps_expB}{A vector (>=2) of replicate numbers that should be considered as second condition's tech reps}

\item{CondReps_expA_name}{Name of the first condition}

\item{CondReps_expB_name}{Name of the second condition}

\item{thr_coverage}{optional parameter, specifies coverage threshold}

\item{minDifference}{minimal difference between AIs in 2 conditions to call them differentially allelically expressed}

\item{save}{If true, save the results to table}
}
\value{
inDF table with 2 additional columns classifying genes into differentially allelically expressed (DAE) and not + figure
}
\description{
A function that gets a table with ref & alt counts per gene/SNP for each replicate
}
\examples{

}
