% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamASE.R
\name{findMAE}
\alias{findMAE}
\title{A function that returns classification based on AI}
\usage{
findMAE(x)
}
\arguments{
\item{x}{Vector with ASE classification per clone}
}
\value{
overall MAE status among clones (by Gendrel classification)
}
\description{
A function that returns classification based on AI
}
\examples{

}
