% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{FitLmIntercept}
\alias{FitLmIntercept}
\title{Fits linear model to logarithmic data and outputs intercept for the model with slope=1/2 restriction}
\usage{
FitLmIntercept(inDF, binNObs, morethan = 10, logoutput = TRUE)
}
\arguments{
\item{inDF}{A dataframe - output of CreateObservedQuantilesDF()}

\item{morethan}{Theshold on gene coverage for lm (default = 10)}

\item{logoutput}{Return log intercept? (default = true)}

\item{N_obs_bin}{Threshold on number of observations per bin}
}
\value{
lm intercept or log2(lm intercept)
}
\description{
Fits linear model to logarithmic data and outputs intercept for the model with slope=1/2 restriction
}
\examples{

}
