% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{GetGatkPipelineSNPTabs}
\alias{GetGatkPipelineSNPTabs}
\title{(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles}
\usage{
GetGatkPipelineSNPTabs(inFiles, nReps)
}
\arguments{
\item{inFiles}{A vector of pathes to files}

\item{nReps}{A vector of numbers, each -- number of replicates in corresponding file}
}
\value{
A technical replicates-concatenated table with SNPs, rows corresponds to SNPs
}
\description{
(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles
}
\examples{

}
