% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGatkPipelineTabs.R
\name{GetGatkPipelineTabs}
\alias{GetGatkPipelineTabs}
\title{GetGatkPipelineTabs}
\usage{
GetGatkPipelineTabs(inFiles, nReps, contigs = vector())
}
\arguments{
\item{inFiles}{A vector of full pathes to files with alelleic counts tables (necessary columns: "ID", pairs of ref and alt counts; optionally, for filtering, "contig")gene names and allelic counts}

\item{nReps}{A vector of numbers, either: (1) each entry is a number of replicates in the corresponding file; (2) for one file only, each entry is a number of replicates corresponding to particular experiment}

\item{contigs}{Parameter defining if the resulting table should be filtered by contig column (preserving only rows corresponding to a given vector), default set to empty vector() and no filtering applied}
}
\value{
A concatenated table with allele counts for all replicates, each row corresponds to a feature ("ID")
}
\description{
Loads the working parts of tables ("ID", allele count 2x columns, "contig" if specified); concatenates (uniting merge) tables from all provided files.
}
