% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMAE_test_CI.R
\name{isMAE_test_CI}
\alias{isMAE_test_CI}
\title{A function that tests for allelic imbalance}
\usage{
isMAE_test_CI(x)
}
\arguments{
\item{x}{Vector two entries: first is the output of CIs test (T if genes are differentially monoallelically expressed), second is absolute AI value}
}
\value{
character (overall MAE status among clones (by Gendrel classification))
}
\description{
A function that tests for allelic imbalance
}
\examples{
isMAE_test_CI(c(TRUE, 0.1, 0.5))
}
