% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogFraction.R
\name{LogFraction}
\alias{LogFraction}
\title{LogFraction}
\usage{
LogFraction(a_up, a_down, j)
}
\arguments{
\item{a_up}{double()}

\item{a_down}{double()}

\item{j}{int()}
}
\value{
log((j + a_up) / (j + a_down))
}
\description{
helper function for computing Gamma
}
