% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanCoverage.R
\name{MeanCoverage}
\alias{MeanCoverage}
\title{MeanCoverage}
\usage{
MeanCoverage(df, reps = NA, thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{df}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{reps}{Optional (default=NA, all replicates), a vector of replicate numbers for which the analysis should be applied}

\item{thr}{Optional (default=NA), threshold on the overall number of counts for a gene to be considered in the analysis}

\item{thrUP}{Optional (default=NA), threshold for max gene coverage (default = NA)}

\item{thrType}{Optional (default = "each", also can be "average" for average coverage on replicates), threshold type}
}
\value{
A table with IDs and calculated mean allelic coverage for given set of replicates
}
\description{
Calculates mean allelic coverage (mat+pat) among given replicates.
}
\examples{
MeanCoverage(allelicCountsTable, reps=c(3,4), thr=8)

}
