% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixBetaBinomialFit.R
\name{MixBetaBinomialFit}
\alias{MixBetaBinomialFit}
\title{MixBetaBinomialFit}
\usage{
MixBetaBinomialFit(initials, coverage, observations)
}
\arguments{
\item{initials}{Initials for EM algm: initials = c(w1, alpha1, alpha2), weight of first component and alphas for both beta-binomial distributions in a mixture}

\item{coverage}{A number, that represents the coverage bin}

\item{observations}{A vector of "maternal counts" in the bin}
}
\value{
Fitted weight of first component and alphas for both beta-binomial distributions in a mixture, plus number of steps.
}
\description{
Fitting beta-binomial mixture distribution of AI in particular bin.
}
