% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{nameColumns}
\alias{nameColumns}
\title{Helper function to quickly rename columns in geneCountTab dataframe}
\usage{
nameColumns(exp_n, rep_n)
}
\arguments{
\item{exp_n}{Experiment number}

\item{rep_n}{Number of replicates for the experiment}
}
\value{
Vector with names
}
\description{
Helper function to quickly rename columns in geneCountTab dataframe
}
\examples{
colnames(geneCountTab)[2:13] <- c(nameColumns(1,2), nameColumns(2,2), nameColumns(3,2))

}
