% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{NumToDoulbledigitChar}
\alias{NumToDoulbledigitChar}
\title{Creates double-digit character from a number or a vector of numbers (up t0 99)}
\usage{
NumToDoulbledigitChar(x)
}
\arguments{
\item{x}{A number from 0 to 99}
}
\value{
Double-digit character (or vector) from a number ('x' or '0x')
}
\description{
Creates double-digit character from a number or a vector of numbers (up t0 99)
}
\examples{

}
