% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBinTestAIAnalysisForConditionNPoint.R
\name{PerformBinTestAIAnalysisForConditionNPoint}
\alias{PerformBinTestAIAnalysisForConditionNPoint}
\title{PerformBinTestAIAnalysisForConditionNPoint}
\usage{
PerformBinTestAIAnalysisForConditionNPoint(
  inDF,
  vectReps,
  pt = 0.5,
  binNObs = 40,
  Q = 0.95,
  fitCovThr = 50,
  EPS = 1.05,
  thr = NA,
  thrUP = NA,
  thrType = "each",
  minDifference = NA
)
}
\arguments{
\item{inDF}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{vectReps}{A vector (>=2) of replicate numbers for which the analysis should be applied}

\item{pt}{Optional (default=0.5), a value to compare with}

\item{binNObs}{Optional (default=40), threshold on number of observations per bin}

\item{Q}{Optional (default=0.95), confidence level, quantile}

\item{fitCovThr}{Optional (default=50), threshold on coverage for genes that will be included in Beta-Bin fitting}

\item{EPS}{Optional (default=1.05), base of exponent for the coverage binning, setting greater base (1.1 or 1.2 or 1.3) would result in fewer number of coverage bins in fitting process, thus will increase the computational speed, but may potentially reduce accuracy}

\item{thr}{Optional (default=NA), threshold on the overall number of counts for a gene to be considered in the analysis}

\item{thrUP}{Optional (default=NA), threshold for max gene coverage (default = NA)}

\item{thrType}{Optional (default = "each", also can be "average" for average coverage on replicates), threshold type}

\item{minDifference}{Optional (default=NA), if specified, one additional column is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to passing the test)}
}
\value{
List of (1) fitted QCC for all combanatorial pairs of replicates ($CC), (2) ComputeCorrConstantsForAllPairsReps() output ($FitDATA), and (3) PerformBinTestAIAnalysisForConditionNPoint_knownCC() output ($Output).
}
\description{
Calculates QCC. Performs Binomial and QCC-corrected binomial tests (with Bonferroni correction) with a given point estimate.
}
