% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/PerformBinTestAIAnalysisForConditionNPointVect.R
\name{PerformBinTestAIAnalysisForConditionNPointVect}
\alias{PerformBinTestAIAnalysisForConditionNPointVect}
\title{PerformBinTestAIAnalysisForConditionNPointVect}
\usage{
PerformBinTestAIAnalysisForConditionNPointVect(
  inDF,
  vectReps,
  ptVect,
  binNObs = 40,
  Q = 0.95,
  fitCovThr = 50,
  EPS = 1.3,
  thr = NA,
  thrUP = NA,
  thrType = "each",
  minDifference = NA
)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{ptVect}{A vector of point to compare with (should be compatible with the order and size of genes vector in table of allelic counts)}

\item{binNObs}{Threshold on number of observations per bin}

\item{Q}{An optional parameter; quantile (for example 0.95, 0.8, etc)}

\item{fitCovThr}{Threshold on coverage for genes that will be included in Beta-Bin fitting}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{minDifference}{if specified, one additional column DAE is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to having non-overlapping CIs)}
}
\value{
List of (1) fitted QCC for all combanatorial pairs of replicates ($CC), (2) ComputeCorrConstantsForAllPairsReps() output ($FitDATA), and (3) PerformBinTestAIAnalysisForConditionNPointVect_knownCC() output ($Output).
}
\description{
Calculate QCC. Perform Binomial and QCC-corrected binomial tests with a given vector of point estimates.
}
