% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBinTestAIAnalysisForTwoConditions.R
\name{PerformBinTestAIAnalysisForTwoConditions}
\alias{PerformBinTestAIAnalysisForTwoConditions}
\title{PerformBinTestAIAnalysisForTwoConditions}
\usage{
PerformBinTestAIAnalysisForTwoConditions(
  inDF,
  vect1CondReps,
  vect2CondReps,
  binNObs = 40,
  fitCovThr = 50,
  Q = 0.95,
  EPS = 1.05,
  thr = NA,
  thrUP = NA,
  thrType = "each",
  minDifference = NA
)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vect1CondReps}{A vector (>=2) of replicate numbers that should be considered as first condition's tech reps}

\item{vect2CondReps}{A vector (>=2) of replicate numbers that should be considered as second condition's tech reps}

\item{binNObs}{Threshold on number of observations per bin}

\item{fitCovThr}{Threshold on coverage for genes that will be included in Beta-Bin fitting}

\item{Q}{An optional parameter; quantile (for example 0.95, 0.8, etc)}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{minDifference}{if specified, one additional column DAE is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to having non-overlapping CIs)}
}
\value{
List of (1) fitted QCC for all combanatorial pairs of replicates for both conditions ($CC), (2) ComputeCorrConstantsForAllPairsReps() output for both conditions ($FitDATA), and (3) PerformBinTestAIAnalysisForTwoConditions_knownCC() output ($Output).
}
\description{
Calculate QCC. Perform differential tests for AI values for two conditions.
}
