% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{PerformBinTestAIAnalysisForTwoConditions_knownCC}
\alias{PerformBinTestAIAnalysisForTwoConditions_knownCC}
\title{Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for condition + point estimate to compare}
\usage{
PerformBinTestAIAnalysisForTwoConditions_knownCC(inDF, vect1CondReps,
  vect2CondReps, vect1CondRepsCombsCC, vect2CondRepsCombsCC, Q = 0.95,
  thr = NA, thrUP = NA, thrType = "each", minDifference = NA)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vect1CondReps}{A vector (>=2) of replicate numbers that should be considered as first condition's tech reps}

\item{vect2CondReps}{A vector (>=2) of replicate numbers that should be considered as second condition's tech reps}

\item{vect1CondRepsCombsCC}{A vector of pairwise-computed correction constants for first condition's tech reps}

\item{vect2CondRepsCombsCC}{A vector of pairwise-computed correction constants for second condition's tech reps}

\item{Q}{An optional parameter; %-quantile (for example 0.95, 0.8, etc)}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{minDifference}{if specified, one additional column DAE is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to having non-overlapping CIs)}
}
\value{
A table of gene names, AIs + CIs, classification into genes demonstrating differential from point estimate AI and those that don't
}
\description{
Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for condition + point estimate to compare
}
\examples{

}
