% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{PerformDAIQuantilesAnalysis}
\alias{PerformDAIQuantilesAnalysis}
\title{Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use}
\usage{
PerformDAIQuantilesAnalysis(inDF, vectReps, condName = "Condition",
  Q = 0.95, EPS = 1.3, thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{condName}{An optional parameter; one-word name for condition}

\item{Q}{An optional parameter; %-quantile (for example 0.95, 0.8, etc)}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{fullOUT}{Set true if you want full output with all computationally-internal dfs}
}
\value{
A table of quantiles in coverage bins
}
\description{
Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use
}
\examples{

}
