% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDesign.R
\name{BuildDesign}
\alias{BuildDesign}
\title{BuildDesign}
\usage{
BuildDesign(experimentNames, techReps, corrConst = NA)
}
\arguments{
\item{experimentNames}{Vector with names of the experiments}

\item{techReps}{Vector with number of technical replicates in each experiment}

\item{corrConst}{Optional, Vector with correction constants for each experiment}
}
\value{
Dataframe with experiments numbered and numbers of columns
}
\description{
Creates a design matrix for the experiment
}
\examples{
BuildDesign(c("clone1", "clone2", "clone3"), c(2, 2, 3), c(1.7, 1.8, mean(1.55, 1.6, 1.57)))

}
