% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCorrConstantFor2Reps.R
\name{ComputeCorrConstantFor2Reps}
\alias{ComputeCorrConstantFor2Reps}
\title{ComputeCorrConstantFor2Reps}
\usage{
ComputeCorrConstantFor2Reps(
  inDF,
  reps,
  binNObs = 40,
  fitCovThr = 50,
  EPS = 1.05,
  thr = NA,
  thrUP = NA,
  thrType = "each"
)
}
\arguments{
\item{inDF}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{reps}{A vector of 2 replicate numbers for which the analysis should be applied}

\item{binNObs}{Optional (default=40), threshold on number of observations per bin}

\item{fitCovThr}{Optional (default=50), threshold on coverage for genes that will be included in Beta-Bin fitting}

\item{EPS}{Optional (default=1.05), base of exponent for the coverage binning, setting greater base (1.1 or 1.2 or 1.3) would result in fewer number of coverage bins in fitting process, thus will increase the computational speed, but may potentially reduce accuracy}

\item{thr}{Optional (default=NA), threshold on the overall number of counts for a gene to be considered in the analysis}

\item{thrUP}{Optional (default=NA), threshold for max gene coverage (default = NA)}

\item{thrType}{Optional (default = "each", also can be "average" for average coverage on replicates), threshold type}
}
\value{
List with (1) fitted QCC ($fittedCC) and (2) a table with proportions of observed to expected quantiles per coverage bin ($QObsExpPropsTable).
}
\description{
Computes QCC for one pair of replicates.
}
