% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsToAI.R
\name{CountsToAI}
\alias{CountsToAI}
\title{CountsToAI}
\usage{
CountsToAI(
  df,
  reps = NA,
  meth = "mergedToProportion",
  thr = NA,
  thrUP = NA,
  thrType = "each"
)
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df)}

\item{meth}{An optional parameter for method to use, either sum(m)/sum(p), or sum(m/p) (default = sum(m)/sum(p))}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
Df with names and mean(mean(m_1,...,m_6))_SNP / mean(mean(m_1+p_1,...,m_6+p6))_SNP
}
\description{
Calculates allelic imbalances from merged counts over given replicates (ai(sum_reps(gene)))
}
\examples{
CountsToAI(allelicCountsTable, reps=c(1,2), thr=10, thrUP=1000)

}
