% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGatkPipelineTabs.R
\name{GetGatkPipelineTabs}
\alias{GetGatkPipelineTabs}
\title{GetGatkPipelineTabs: Concatenate vertically (uniting merge) tables from inFiles.}
\usage{
GetGatkPipelineTabs(inFiles, nReps, contigs = vector())
}
\arguments{
\item{inFiles}{A vector of full pathes to files with gene names and allelic counts}

\item{nReps}{A vector of numbers, each entry is a number of replicates in the corresponding file}

\item{contigs}{Parameter defining if the resulting table should be filtered by contig column preserving only row corresponding to a given vector, default set to empty vector() and no filtering applied}
}
\value{
A concatenated table with means/counts, each row corresponds to a gene
}
\description{
GetGatkPipelineTabs: Concatenate vertically (uniting merge) tables from inFiles.
}
