% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{GetGatkPipelineTabs}
\alias{GetGatkPipelineTabs}
\title{(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles.}
\usage{
GetGatkPipelineTabs(inFiles, nReps, multiple = TRUE, chrom = F)
}
\arguments{
\item{inFiles}{A vector of full pathes to files with gene names and allelic counts}

\item{nReps}{A vector of numbers, each entry is a number of replicates in the corresponding file}

\item{multiple}{Parameter defining if multiple input files are used, default set to TRUE}

\item{chrom}{Parameter defining if the resulting table includes chromosome column, default set to FALSE}
}
\value{
A concatenated table with means/counts, each row corresponds to a gene
}
\description{
(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles.
}
\examples{

}
