% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{LogFraction}
\alias{LogFraction}
\title{Input: 3 numbers (double, double, int)}
\usage{
LogFraction(a_up, a_down, j)
}
\arguments{
\item{a_up}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{a_down}{A vector (2) of replicate numbers that should be considered}
}
\value{
log((j + a_up) / (j + a_down))
}
\description{
Input: 3 numbers (double, double, int)
}
\examples{

}
