% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanCoverage.R
\name{MeanCoverage}
\alias{MeanCoverage}
\title{MeanCoverage}
\usage{
MeanCoverage(df, reps = NA, thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns.}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df).}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
Df with names and Mean coverage mat+pat among given replicates.
}
\description{
Calculates mean coverage mat+pat among given replicates.
}
\examples{
MeanCoverage(allelicCountsTable, reps=c(3,4), thr=8)

}
