% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSumCounts.R
\name{MergeSumCounts}
\alias{MergeSumCounts}
\title{MergeSumCounts}
\usage{
MergeSumCounts(df, reps = NA, thr = NA, thrUP = NA, thrType = "each")
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns.}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df).}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
Table with names and sums of mat and pat coverages among given replicates.
}
\description{
Creates a table of sums of maternal and paternal count for given replicates.
}
\examples{
MeanCoverage(Abl1_allele_counts, reps=c(1,2))

}
