% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{MixBetaBinomialFitStep}
\alias{MixBetaBinomialFitStep}
\title{Input: 3 numbers (double, double, int)}
\usage{
MixBetaBinomialFitStep(initials_old, coverage, observations)
}
\arguments{
\item{initials_old}{Initials for EM step: initials = c(w1, alpha1, alpha2), weight of first component and alphas for both beta-binomial distributions in a mixture}

\item{coverage}{A number, that reflects higher boundary of the coerage bin}

\item{observations}{A vector of maternal counts in the bin}
}
\value{
Re-fitted initials for next EM step.
}
\description{
Input: 3 numbers (double, double, int)
}
\examples{

}
