% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameColumns.R
\name{NameColumns}
\alias{NameColumns}
\title{NameColumns}
\usage{
NameColumns(exp_n, rep_n)
}
\arguments{
\item{exp_n}{Experiment number}

\item{rep_n}{Number of replicates for the experiment}
}
\value{
Vector with names
}
\description{
Helper function to quickly rename columns in geneCountTab dataframe
}
\examples{
colnames(Abl_1and2_allelic_counts)[2:9] <- c(NameColumns(1,2), NameColumns(2,2))

}
