% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformAIAnalysis_CC.R
\name{PerformBinTestAIAnalysisForConditionNPoint}
\alias{PerformBinTestAIAnalysisForConditionNPoint}
\title{Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for condition + point estimate to compare}
\usage{
PerformBinTestAIAnalysisForConditionNPoint(inDF, vectReps, pt = 0.5,
  binNObs = 40, Q = 0.95, EPS = 1.3, thr = NA, thrUP = NA,
  thrType = "each", minDifference = NA)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{pt}{A point to compare with}

\item{binNObs}{Threshold on number of observations per bin}

\item{Q}{An optional parameter; %-quantile (for example 0.95, 0.8, etc)}

\item{EPS}{An optional parameter to set a log window for coverage binning}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{minDifference}{if specified, one additional column DAE is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to having non-overlapping CIs)}
}
\value{
A table of gene names, AIs + CIs, classification into genes demonstrating differential from point estimate AI and those that don't
}
\description{
Input: data frame with gene names and counts (reference and alternative) + numbers of replicates to use for condition + point estimate to compare
}
\examples{

}
