% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/PerformBinTestAIAnalysisForConditionNPointVect_knownCC.R
\name{PerformBinTestAIAnalysisForConditionNPointVect_knownCC}
\alias{PerformBinTestAIAnalysisForConditionNPointVect_knownCC}
\title{PerformBinTestAIAnalysisForConditionNPointVect_knownCC}
\usage{
PerformBinTestAIAnalysisForConditionNPointVect_knownCC(
  inDF,
  vectReps,
  vectRepsCombsCC,
  ptVect,
  Q = 0.95,
  thr = NA,
  thrUP = NA,
  thrType = "each",
  minDifference = NA
)
}
\arguments{
\item{inDF}{A table with ref & alt counts per gene/SNP for each replicate plus the first column with gene/SNP names}

\item{vectReps}{A vector (>=2) of replicate numbers that should be considered as tech reps}

\item{vectRepsCombsCC}{A vector of pairwise-computed correction constants for given replicates}

\item{ptVect}{A vector of point to compare with (should be compatible with the order and size of genes vector in table of allelic counts)}

\item{Q}{An optional parameter; quantile (for example 0.95, 0.8, etc)}

\item{thr}{An optional parameter; threshold on the overall number of counts (in all replicates combined) for a gene to be considered}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}

\item{minDifference}{if specified, one additional column DAE is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to having non-overlapping CIs)}
}
\value{
A table of gene names, AIs + CIs, classification into genes demonstrating signifficant difference (TRUE) from corresponding point estimate AI and those that don't (FALSE)
}
\description{
Perform Binomial and QCC-corrected binomial tests with a given vector of point estimates, for given QCC.
}
