% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/PerformBinTestAIAnalysisForTwoConditions_knownCC.R
\name{PerformBinTestAIAnalysisForTwoConditions_knownCC}
\alias{PerformBinTestAIAnalysisForTwoConditions_knownCC}
\title{PerformBinTestAIAnalysisForTwoConditions_knownCC}
\usage{
PerformBinTestAIAnalysisForTwoConditions_knownCC(
  inDF,
  vect1CondReps,
  vect2CondReps,
  vect1CondRepsCombsCC,
  vect2CondRepsCombsCC,
  Q = 0.95,
  thr = NA,
  thrUP = NA,
  thrType = "each",
  minDifference = NA
)
}
\arguments{
\item{inDF}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{vect1CondReps}{A vector (>=2) of replicate numbers that should be considered as first condition's tech reps}

\item{vect2CondReps}{A vector (>=2) of replicate numbers that should be considered as second condition's tech reps}

\item{vect1CondRepsCombsCC}{A vector of pairwise-computed correction constants for first condition's tech reps (QCC=1 is no correction)}

\item{vect2CondRepsCombsCC}{A vector of pairwise-computed correction constants for second condition's tech reps (QCC=1 is no correction)}

\item{Q}{Optional (default=0.95), confidence level, quantile}

\item{thr}{Optional (default=NA), threshold on the overall number of counts for a gene to be considered in the analysis}

\item{thrUP}{Optional (default=NA), threshold for max gene coverage (default = NA)}

\item{thrType}{Optional (default = "each", also can be "average" for average coverage on replicates), threshold type}

\item{minDifference}{Optional (default=NA), if specified, one additional column is added to the output (T/F depending if the gene changed AI expression more than minDifference in addition to passing the test)}
}
\value{
A table of gene names, AIs + CIs for both conditions, p-values for both non-corrected (BT..) and QCC corrected (BT_CC..) differential tests, classification into genes demonstrating signifficant difference (TRUE) of AI estimates in two conditions, and those that don't (FALSE).
}
\description{
Performs differential tests (with Bonferroni correction) for AI values for two conditions, for given QCC.
}
