% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE_functions.R
\name{ThresholdingCounts}
\alias{ThresholdingCounts}
\title{Takes table with gene names and counts and returns table, where all genes that under threshold have NA coverage. Can be restricted to particular reps.}
\usage{
ThresholdingCounts(df, reps = NA, thr = NA, thrUP = NA,
  thrType = "each")
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df)}

\item{thr}{An optional parameter for a threshold on gene coverage (default = NA)}

\item{thrUP}{An optional parameter for a threshold for max gene coverage (default = NA)}

\item{thrType}{An optional parameter for threshold type (default = "each", also can be "average" coverage on replicates)}
}
\value{
Table with masked with NA undercovered genes
}
\description{
Takes table with gene names and counts and returns table, where all genes that under threshold have NA coverage. Can be restricted to particular reps.
}
\examples{

}
